Remove GitLab on-prem
Since SSO appears to be included as a basic feature for the on-prem version of GitLab, it doesn't seem like it belongs on this list.  Additionally, according the the GitLab feature list, it's not actually SSO but ldap sync and auth.  Which does accomplish some of the same goals but does not provide the same user experience as true SSO.