remove sumo logic as they have seen the light
It took me 2 years but I finally convinced them to move it into the standard tiers of pricing. The only tier that does not support it is free tier at 500mb but I think thats a fair stance to take if you are ingesting less than 500mb a day you likely dont have a team which is why SSO is kinda a baseline requirement for orgs.

Signed-off-by: Ben Abrams <me@benabrams.it>