Rename url parameter to vendor_url
In #254, we migrated to using a Jekyll collection to keep each vendor contained to its own yaml file. As part of this,
Jekyll injects its own `url` parameter into the resulting document. This has caused all homepage vendor links to be
broken.

To address this, vendors will now use `vendor_url` to note the vendor homepage.